One third of the exercises feature is almost completed Score Date now has a score editor that allows users to create any kind of sequence, for any game.
It turned out that MusiSync (not MusicSync...) is a great choice when it
comes to antialiasing and flexibility. You just need to be a capable
developer to be able to use it. Someone who doesn't think so could
always revert back to some crappy gif clips and years of latency.